refactor spot feed, add vol feed

+++ src/interfaces/IVolFeed.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.18;
 
 interface IVolFeed {
-  function getVol(uint128 strike, uint128 expiry) external view returns (uint128 vol, uint64 confidence);
+  function getVol(uint128 strike, uint64 expiry) external view returns (uint128 vol, uint64 confidence);
 
   ////////////
   // Events //
